Initiating Conways Game of Life VBA
Two buttons will have to be created to use this code. The first named CommandButton1 will allow you to select  cell and will give you the 'cell count' as in all the green cells in the 3 x 3 block including the cell that has been selected.

The second is CommandButton2, and that will start the code that runs Conway's Game of life. This script is set to run the application over and over again (30) times and can be adjusted by modifying the loop "For i = 0 to 30"